LU-11527 tests: fix sanity 270a for ARM
Fix sanity test_270a to work with ARM so that it does O_DIRECT
writes with a multiple of the local PAGE_SIZE (65536 bytes)
instead of 102400 bytes (which works fine on x64, but not ARM).

Generalize the test to use variables instead of precomputed values
so that it is clear why the particular values are used.

Test-Parameters: trivial envdefinitions=ONLY=270a
Test-Parameters: mdtfilesystemtype=zfs ostfilesystemtype=zfs testlist=sanity envdefinitions=ONLY=270a
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Ic2a32215dc1423b0bbf1e5188a5e85420a3ebbe5
Reviewed-on: https://review.whamcloud.com/33633
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-by: Jian Yu <yujian@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>