LU-11597 tests: fix O_DIRECT test usage for ARM
Fix cases where dd [io]flag=direct is used but the blocksize is
not a multiple of the client PAGE_SIZE, which is required for
Lustre.

Add a global PAGE_SIZE variable in init_test_env() and use that
in the tests, rather than having multiple local functions/calls
to re-determine the PAGE_SIZE value, which will not change for
the duration of the test.

Keep the get_page_size() function to initialize the global
PAGE_SIZE value and for the few cases where tests use the server
PAGE_SIZE (e.g. obdecho testing).

Test-Parameters: trivial
Test-Parameters: mdtfilesystemtype=zfs ostfilesystemtype=zfs testlist=sanity
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I0bb9efbcc743f6508180c8496fc50a67363ebbe5
Reviewed-on: https://review.whamcloud.com/33636
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-by: Jian Yu <yujian@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>