LU-11566 utils: improve usage/docs for lctl llog commands
Improve the usage message and man pages for the llog_print,
llog_info, llog_catlist, and llog_cancel sub-commands.  Move
them out of the obsolete section of the lctl usage message.

Reorder some man pages to be in alphabetical order in Makefile.

Add named options to the various commands to make them
consistent with each other.  The --catalog option is not
required for ease of use, but is kept for compatibility
with the previous llog_cancel interface.

The llog_cancel --log_id option is removed from the usage
message and man page, since the ability to cancel individual
records from MDT recovery logs is currently not implemented
(no IOC_LLOG_CANCEL handler in mdt_iocontrol()).  There is
also no IOC_LLOG_PRINT method for the MDS either, so these
commands are mostly useful only for the MGS configuration
logs at this time.

Add a test case for llog_cancel.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I66c95c289f161370896e6764da4fc2f5803ebbe5
Reviewed-on: https://review.whamcloud.com/34004
Tested-by: Jenkins
Reviewed-by: Joseph Gmitter <jgmitter@whamcloud.com>
Reviewed-by: Ben Evans <bevans@cray.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>