LU-11329 misc: reorder Git .mailmap file
If there are multiple remappings of the same person (e.g. CFS to
Sun to Oracle to Whamcloud to Intel to Whamcloud2), the entries
in the .mailmap file need to be in chronological order, so that
older entries are mapped to the next email, which is mapped to
the next one, etc.  Otherwise, only a single remapping is done.

Add additional mapping for Patrick Farrell.

Test-Parameters: trivial
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I6d7f2977da5c24707e3d54fddfdabf94c28e3f0b
Reviewed-on: https://review.whamcloud.com/34109
Reviewed-by: Patrick Farrell <pfarrell@whamcloud.com>
Tested-by: Jenkins
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Peter Jones <pjones@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>