LU-11568 ldlm: Remove use of SLAB_DESTROY_BY_RCU for ldlm lock slab
Whatever it was doing does not appear to be necessary anymore
as evidenced with newer kernels where the define was removed,
but disabled in Lustre instead.
Another important reason to remove it is because rhel7.3+ seems
to have broken this RCU functionality and leads to frequent use
after frees.

This patch is back-port from:
Lustre-commit: 82d014e71e14671e876055851a0d37e98b4cc079
Lustre-change: https://review.whamcloud.com/34147

Change-Id: I50991b9daf4ef06b24cb65d7a04a5e9b86706d36
Signed-off-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-on: https://review.whamcloud.com/34434
Tested-by: Jenkins