LU-11010 tests: remove calls to return after skip()
The skip() routine now contains a call to exit. All calls
to skip() and skip_env() should be reviewed and calls to
return that followed skip should be removed.

A problem with the skip message not being printed is corrected.

This patch is partial backport (sanity-pfl is not modified) from:
Lustre-commit: ea76de56fccfdc6f8bd0087adc36c19257e414d3
Lustre-change: https://review.whamcloud.com/32346

Signed-off-by: James Nunez <jnunez@whamcloud.com>
Change-Id: I1a52e9bd79a71de4ab4c0cea9c569f379115a603
Reviewed-on: https://review.whamcloud.com/34199
Tested-by: Jenkins
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Wei Liu <sarah@whamcloud.com>